extract proxy configuration for re-use in rest api v2 client
